[refactor] 커스텀 캘린더 공통 컴포넌트화
